+++ Exodus/src/utopia/exodus/model/stored/auth/TokenScopeLink.scala
@@ -1,6 +1,7 @@
 package utopia.exodus.model.stored.auth
 
 import utopia.exodus.database.access.single.auth.DbSingleTokenScopeLink
+import utopia.exodus.model.combined.auth.TokenScope
 import utopia.exodus.model.partial.auth.TokenScopeLinkData
 import utopia.vault.model.template.StoredModelConvertible
 
@@ -20,5 +21,14 @@ case class TokenScopeLink(id: Int, data: TokenScopeLinkData)
 	  * An access point to this token scope link in the database
 	  */
 	def access = DbSingleTokenScopeLink(id)
+	
+	
+	// OTHER    ---------------------
+	
+	/**
+	  * @param scope Scope information to attach
+	  * @return A copy of this link with that infor attached
+	  */
+	def withScopeInfo(scope: Scope) = TokenScope(scope, this)
 }